refactoring: creates articles.models and topics.models and put all the associated functions in there
