+++ models/articles.models.js
@@ -21,3 +21,11 @@ exports.setArticles = () => {
     return rows;
   });
 };
+
+exports.selectArticleById = (article_id) => {
+  return db
+    .query("SELECT * FROM articles WHERE article_id = $1", [article_id])
+    .then(({ rows }) => {
+      return rows[0];
+    });
+};
