--- models/selectArticleById.models.js
@@ -1,9 +1,0 @@
-const db = require("../db/connection");
-
-exports.selectArticleById = (article_id) => {
-  return db
-    .query("SELECT * FROM articles WHERE article_id = $1", [article_id])
-    .then(({ rows }) => {
-      return rows[0];
-    });
-};
